feat(surface): unify MCP+REST+CLI+OpenAPI behind one command registry (#2)
One declarative commands.py registry generates the MCP tools (via bind_vault,
byte-identical schemas — proven by tests/test_mcp_schema_fidelity.py against a
fixture captured from the pre-refactor server), the REST facade + OpenAPI, and a
new kb/kb-mcp CLI (reads + writes, Endstate-style --json envelope + exit codes).
server.py -2019 lines (wrappers -> registry leaves, AST-extracted verbatim).
Exceptions {note, mint_upload_token, mint_download_token} stay hand-registered.
One behavior change: the 9 REST routes now return {success,data|error}.

Code-reviewed (APPROVE, 0 blockers; independently verified byte-identical +
23/23 bodies AST-identical); 785 passed; OAuth/upload/download/file-watcher
intact. LOW follow-ups tracked in tasks.md sec 9.
