+++ pyproject.toml
@@ -85,6 +85,13 @@ dev = [
     "httpx>=0.27",
 ]
 
+[project.scripts]
+# `kb find "…"` is the daily command; `kb-mcp` is the namespaced alias (and still
+# carries the admin subcommands: init / install-skill / serve / …). `python -m
+# kb_mcp` keeps working too — all three route to the same entry point.
+kb = "kb_mcp.__main__:main"
+kb-mcp = "kb_mcp.__main__:main"
+
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
