+++ src/kb_mcp/commands.py
@@ -1,0 +1,1938 @@
+"""Single declarative command registry — the genuine source of truth for every
+surface (MCP tools, the REST facade, the OpenAPI document, and the CLI).
+
+Each operation is one `Command`: its canonical name, the leaf callable
+`leaf(vault_root, **kwargs)` (the former per-surface wrapper body, lifted to
+module level so it can be shared), declarative `Param` specs (drive REST
+coercion + CLI argparse + OpenAPI), the set of surfaces it is exposed on, and the
+full description Claude reads (the leaf's own docstring).
+
+MCP tools are generated via `bind_vault`, which presents each leaf's signature
+(minus the injected `vault_root` / `source_schema`) and its docstring to FastMCP
+exactly as a hand-written wrapper would — so the generated tool's input-schema and
+description are byte-identical to the pre-registry tool (pinned by
+`tests/test_mcp_schema_fidelity.py`). Any tool whose schema cannot be reproduced
+cleanly (the env-bound `mint_*` tools) — or that needs a per-vault description
+(`note`'s live project-key hint) — stays hand-registered in `server.py` and is
+named in `HAND_REGISTERED_EXCEPTIONS`.
+"""
+
+from __future__ import annotations
+
+import inspect
+import types
+import typing
+from collections.abc import Callable
+from dataclasses import dataclass
+from pathlib import Path
+
+from . import add as add_module
+from . import append_to_file as append_to_file_module
+from . import audit as audit_module
+from . import audit_fix as audit_fix_module
+from . import compile_proposal as compile_proposal_module
+from . import corpus_aware as corpus_aware_module
+from . import create_directory as create_directory_module
+from . import create_file as create_file_module
+from . import delete_directory as delete_directory_module
+from . import delete_file as delete_file_module
+from . import edit as edit_module
+from . import find as find_module
+from . import get_frontmatter as get_frontmatter_module
+from . import get_page as get_page_module
+from . import link as link_module
+from . import list_directory as list_directory_module
+from . import list_inbound_links as list_inbound_links_module
+from . import list_trash as list_trash_module
+from . import move_file as move_file_module
+from . import multi_edit as multi_edit_module
+from . import note as note_module
+from . import preserve as preserve_module
+from . import provenance as provenance_module
+from . import query_data as query_data_module
+from . import query_log, vault
+from . import reconcile as reconcile_module
+from . import recover_from_trash as recover_from_trash_module
+from . import replace as replace_module
+from . import set_frontmatter_field as set_frontmatter_field_module
+from . import set_take as set_take_module
+from .vault import (
+    VaultPathError,
+    find_body_wikilinks,
+    resolve_under_vault,
+)
+
+# Text-write ops → the argument field(s) whose value must not be a base64 binary
+# blob. The model pays for those characters as output tokens before the request
+# arrives, so they are rejected at every write boundary (MCP middleware + REST
+# coercion) and the caller is pointed at /upload.
+GUARDED_WRITE_FIELDS: dict[str, tuple[str, ...]] = {
+    "add": ("content",),
+    "note": ("content",),
+    "edit": ("new_body", "new_string"),
+    "replace": ("content",),
+    "create_file": ("content",),
+    "append_to_file": ("content",),
+    "preserve": ("content",),
+}
+
+
+# --------------------------------------------------------------------------- #
+# Registry dataclasses
+# --------------------------------------------------------------------------- #
+@dataclass(frozen=True)
+class Param:
+    """One operation parameter, surface-agnostic.
+
+    `type` is a coercion tag ("str" | "int" | "bool" | "list[str]" | "dict" |
+    "json") used by the REST body coercion, the CLI argparse builder, and the
+    OpenAPI schema — NOT by MCP (MCP derives its schema from the leaf's real
+    signature via `bind_vault`). `cli_positional` makes the CLI take it as a
+    positional arg (at most one per command).
+    """
+
+    name: str
+    type: str = "str"
+    required: bool = False
+    help: str = ""
+    cli_positional: bool = False
+
+
+@dataclass(frozen=True)
+class Command:
+    name: str  # canonical op name — identical across all surfaces
+    leaf: Callable  # leaf(vault_root, **kwargs); for `add`, leaf(vault_root, source_schema, **kwargs)
+    params: tuple[Param, ...]
+    surfaces: frozenset  # subset of {"mcp", "rest", "cli"}
+    tier: int = 1
+    cli_writes: bool = False  # marks a vault-mutating op (CLI grouping / future confirms)
+    needs_schema: bool = False  # leaf takes source_schema as its 2nd injected arg (`add`)
+    description: str = ""
+
+    @property
+    def doc(self) -> str:
+        """The full description Claude reads — the leaf's own docstring."""
+        return self.description or (self.leaf.__doc__ or "")
+
+    @property
+    def guarded_fields(self) -> tuple[str, ...]:
+        """Text fields whose value must not be a base64 binary blob."""
+        return GUARDED_WRITE_FIELDS.get(self.name, ())
+
+
+# --------------------------------------------------------------------------- #
+# bind_vault — present a leaf as an MCP-introspectable callable
+# --------------------------------------------------------------------------- #
+def bind_vault(
+    leaf: Callable,
+    *injected: object,
+    name: str | None = None,
+    description: str | None = None,
+) -> Callable:
+    """Return a callable FastMCP introspects exactly like a hand-written wrapper.
+
+    The returned wrapper:
+    - has `__signature__` = `leaf`'s signature minus the leading `injected` params
+      (always `vault_root`, plus `source_schema` for `add`),
+    - has `__annotations__` resolved against `leaf`'s own module (so string
+      annotations from `from __future__ import annotations` become real types),
+    - has `__doc__` = `description` (the registry/leaf docstring), and
+    - calls `leaf(*injected, **kwargs)`.
+
+    FastMCP builds the tool's input-schema from that signature + the parsed
+    docstring, so the generated tool is byte-identical to the original wrapper.
+    """
+    sig = inspect.signature(leaf)
+    params = list(sig.parameters.values())
+    visible = params[len(injected):]
+
+    # Resolve string annotations (PEP 563) against the leaf's real globals so the
+    # schema doesn't depend on this module's namespace.
+    try:
+        resolved = typing.get_type_hints(leaf)
+    except Exception:  # noqa: BLE001 — fall back to whatever inspect carries
+        resolved = {}
+
+    visible = [p.replace(annotation=resolved.get(p.name, p.annotation)) for p in visible]
+    new_sig = sig.replace(parameters=visible)
+
+    def wrapper(**kwargs):
+        return leaf(*injected, **kwargs)
+
+    wrapper.__signature__ = new_sig  # type: ignore[attr-defined]
+    wrapper.__name__ = name or leaf.__name__
+    wrapper.__qualname__ = wrapper.__name__
+    wrapper.__doc__ = description if description is not None else leaf.__doc__
+    ann = {p.name: p.annotation for p in visible if p.annotation is not inspect.Parameter.empty}
+    if "return" in resolved:
+        ann["return"] = resolved["return"]
+    wrapper.__annotations__ = ann
+    return wrapper
+
+
+# --------------------------------------------------------------------------- #
+# Shared helper used by the `get` / suggest leaves
+# --------------------------------------------------------------------------- #
+def _link_summary(vault_root: Path, rel_path: str, body: str) -> dict:
+    """Inbound + outbound wikilink summary for the `get(links=True)` option.
+
+    Inbound reuses `vault.find_inbound_wikilinks` (the same matcher
+    `list_inbound_links` uses); outbound scans the body's wikilinks (code spans
+    skipped), de-duped and order-preserved.
+    """
+    inbound = (
+        [m.as_dict() for m in vault.find_inbound_wikilinks(vault_root, rel_path)]
+        if rel_path
+        else []
+    )
+    outbound: list[str] = []
+    seen: set[str] = set()
+    for m in find_body_wikilinks(body):
+        target = m.group(0)[2:-2].split("|", 1)[0].split("#", 1)[0].strip()
+        if target and target not in seen:
+            seen.add(target)
+            outbound.append(target)
+    return {"inbound": inbound, "outbound": outbound}
+
+
+# ----- op-leaves: the former per-surface wrapper bodies (vault_root injected) -----
+# Extracted verbatim from server.py's build_server; their docstrings ARE the tool
+# descriptions Claude reads (byte-pinned by tests/test_mcp_schema_fidelity.py).
+
+
+def op_find(
+    vault_root: Path,
+    query: str = "",
+    types: list[str] | None = None,
+    projects: list[str] | None = None,
+    tags: list[str] | None = None,
+    file_types: list[str] | None = None,
+    exclude_file_types: list[str] | None = None,
+    limit: int = 15,
+    scope: str = "kb",
+    mode: str = "hybrid",
+    graph: bool = True,
+    rerank: bool = False,
+    prefer_compiled: bool = True,
+    prefer_active: bool = True,
+) -> list[dict]:
+    """Search / find / look up / query / retrieve / recall pages in the Knowledge Base (KB vault): notes, sources, insights, failures, patterns, experiments, entities. Hybrid semantic + keyword search, read-only. Filters are AND'd; tag/project lists are OR'd within.
+
+    Args:
+        query: Free-text search string. In "hybrid"/"vector" mode it's
+            embedded with bge-base for semantic recall. In "keyword" mode
+            it's tokenized on whitespace and every token must appear in
+            title or body (any order) — `contract employment` matches a
+            page about "employment contract". Empty string always falls
+            back to "most-recent filtered" behaviour regardless of mode.
+        types: Filter to these page types (source, research-note, insight, failure, pattern, experiment, production-log, entity).
+        projects: Filter to pages whose `project` or `projects:` includes any of these keys.
+        tags: Filter to pages whose `tags:` includes any of these (case-insensitive).
+        file_types: Scope results to these artifact kinds — note, pdf, image,
+            audio, video, csv, json, tsv. A binary surfaces under its media
+            kind (pdf/image/...); a data file under its dataset card's format
+            (csv/json). Omit to return ALL kinds (the default — search never
+            hides a type unless you ask).
+        exclude_file_types: Drop these kinds from results (same vocabulary).
+        limit: Max hits to return. Default 15, hard cap 100.
+        scope: "kb" (default) searches Knowledge Base/ first and
+            AUTO-WIDENS to the whole vault when the KB doesn't fill
+            `limit` — so content in sibling folders (Tracking/,
+            Reference/, Finance/, ... and curated, read-only trees kept
+            outside Knowledge Base/) is never silently invisible. Widened
+            hits carry `outside_kb: true`. "vault" always walks the
+            whole vault. "kb-only" is the strict opt-out: KB only,
+            never widens. Outside-KB recall is BM25/keyword (the
+            vector sidecar is KB-scoped), with a relaxed gate so terse
+            files (e.g. a numbers-heavy tracker) surface on a partial
+            token match. `_Schema/`, `_trash/`, `_attachments/`, and
+            `.obsidian/` are excluded under every scope. NOTE: an
+            empty result means "not found in what I searched," NOT
+            "doesn't exist" — say so, and try "vault" before
+            concluding absence.
+        mode: Ranker. "hybrid" (default) fuses BM25 + local vector
+            embeddings via reciprocal rank fusion — best recall on
+            natural-language queries. "keyword" preserves the original
+            case-insensitive substring matching, sorted by `updated:`.
+            "vector" is vector-only (testing aid). BM25 corpus is
+            Snowball-stemmed so "regulation" reaches pages with
+            "regulator"; keyword mode stays strict-substring. If the
+            embedding sidecar hasn't been built yet, hybrid degrades
+            to BM25-only; run `audit_fix(rebuild_embeddings=true)` to
+            populate it.
+        graph: When true (default) and mode is hybrid/vector, outbound
+            wikilinks of top BM25/vector candidates contribute a third
+            ranking — surfaces 1-hop neighbours of strong matches.
+        rerank: When true (off by default), runs the top fused
+            candidates through bge-reranker-base (a CrossEncoder) for
+            higher-precision ordering. Adds ~50ms/candidate; useful
+            when ambiguous queries float topically-off vector matches
+            to the top.
+        prefer_compiled: When true (default), applies a small boost to
+            compiled types (insight, pattern, failure, research-note,
+            entity) and a small penalty to raw `source` after fusion
+            AND rerank. Reflects the KB's epistemic hierarchy. Set
+            false to retrieve raw source discussion verbatim (e.g.
+            "what did I capture from Dr. X").
+        prefer_active: When true (default), soft-demotes `status:
+            superseded` pages so a replaced conclusion can't outrank the
+            page that superseded it. The tombstone stays findable and its
+            hit still carries `status` + `superseded_by` (the forward
+            pointer) so you can see it's superseded. Set false to rank a
+            superseded page on its content alone (e.g. "what did I used to
+            think about X").
+
+    Returns:
+        List of {path, type, scope, title, updated, excerpt[, outside_kb]
+        [, status][, superseded_by][, signals]}. `outside_kb: true` is
+        present only on hits the "kb" auto-widen pulled from beyond
+        Knowledge Base/ (the `path` also shows the sibling folder).
+        `status` + `superseded_by` appear only when a hit is NOT plain
+        `active` — i.e. a superseded tombstone (or draft) — so you can tell
+        it from a live conclusion and follow `superseded_by` to the replacement.
+        In hybrid mode `excerpt` shows the best-matching chunk; in
+        keyword mode it's a snippet anchored to the literal query
+        match. `signals` (hybrid/vector only) carries per-ranker
+        position: {bm25_rank?, vector_rank?, vector_score?, graph_hop?,
+        graph_in_degree?, rerank_score?}. `graph_in_degree` is the
+        number of top-N seeds whose body wikilinks to this hit —
+        independent of graph_hop, which only fires for graph-only
+        results.
+    """
+    hits = find_module.find(
+        vault_root,
+        query=query,
+        types=types,
+        projects=projects,
+        tags=tags,
+        file_types=file_types,
+        exclude_file_types=exclude_file_types,
+        limit=limit,
+        scope=scope,
+        mode=mode,
+        graph=graph,
+        rerank=rerank,
+        prefer_compiled=prefer_compiled,
+        prefer_active=prefer_active,
+    )
+    # Durable structured log → feeds the offline retrieval feedback loop.
+    # Best-effort; never affects the returned result.
+    query_log.log_find_call(
+        query=query, mode=mode, scope=scope,
+        types=types, projects=projects, tags=tags,
+        limit=limit, rerank=rerank, prefer_compiled=prefer_compiled,
+        graph=graph, hits=hits,
+    )
+    return [h.as_dict() for h in hits]
+
+
+def op_suggest_links(
+    vault_root: Path,
+    path: str | None = None,
+    draft_title: str | None = None,
+    draft_body: str | None = None,
+    limit: int = 8,
+    scope: str = "kb",
+) -> list[dict]:
+    """Suggest existing KB pages a note should link to. Read-only.
+
+    Closes the corpus-blind-write gap: surfaces the related prior work a
+    draft (or an existing page) should connect to, so the graph gets denser
+    with every write instead of just bigger. For link suggestions only — it
+    reuses the same hybrid ranker as `find`, prefers well-connected hubs, and excludes
+    the page itself plus anything it already links. Suggestions are
+    non-binding: YOU decide which to wire in (e.g. via a follow-up `edit`).
+
+    Two call shapes:
+    - `path`: suggest links for an EXISTING page (densify it retroactively).
+      Same path conventions as `get`/`find`.
+    - `draft_title` + `draft_body`: suggest links for a note you're about to
+      create, BEFORE calling `note` — so you can cite/connect on first write.
+
+    Args:
+        path: Existing page to suggest links for. Mutually exclusive with
+            the draft_* args.
+        draft_title: Title of a not-yet-written note.
+        draft_body: Body (markdown) of a not-yet-written note. Wikilinks
+            already present in it are treated as "already linked" and excluded.
+        limit: Max suggestions (default 8).
+        scope: "kb" (default) or "vault" — same meaning as `find`.
+
+    Returns:
+        List of {path, title, type, why, excerpt}, best-first. `why`
+        explains the match (e.g. "semantic #2, 4 shared link(s) (hub)").
+        Empty list if nothing relevant or the draft/page is empty.
+
+    Errors:
+        INVALID_SUGGEST (neither path nor draft supplied); plus get-style
+        path errors (NOT_FOUND, INVALID_PATH) when `path` doesn't resolve.
+    """
+    if path:
+        try:
+            gp = get_page_module.get_page(vault_root, path=path)
+        except get_page_module.GetError as e:
+            raise ValueError(f"{e.code}: {e.reason}") from e
+        page = find_module._CACHE.get(vault_root / gp.path, vault_root)
+        if page is None:
+            raise ValueError(f"UNREADABLE: could not parse {gp.path}")
+        existing_links = set(
+            find_module._outbound_wikilink_paths(page, vault_root)
+        )
+        suggestions = corpus_aware_module.suggest_related(
+            vault_root, title=page.title, body=page.body,
+            self_path=page.rel_path, existing_links=existing_links,
+            limit=limit, scope=scope,
+        )
+    elif draft_title or draft_body:
+        body = draft_body or ""
+        existing_links = set()
+        for m in find_body_wikilinks(body):
+            inner = m.group(0)[2:-2].split("|", 1)[0].split("#", 1)[0].strip()
+            if inner:
+                existing_links.add(inner)
+        suggestions = corpus_aware_module.suggest_related(
+            vault_root, title=draft_title or "", body=body,
+            self_path=None, existing_links=existing_links,
+            limit=limit, scope=scope,
+        )
+    else:
+        raise ValueError(
+            "INVALID_SUGGEST: provide either `path` (existing page) or "
+            "`draft_title`/`draft_body` (a note you're about to write)"
+        )
+    return [s.as_dict() for s in suggestions]
+
+
+def op_add(
+    vault_root: Path,
+    source_schema: object,  # SourceSchema; injected + stripped, so kept import-free here
+    content: str,
+    source_type: str,
+    title: str,
+    url: str | None = None,
+    tags: list[str] | None = None,
+    why_captured: str | None = None,
+) -> dict:
+    """Capture raw content as an immutable source page in the Knowledge Base.
+
+    Writes a frontmatter-compliant page to Sources/<Type>/YYYY-MM-DD-<slug>.md
+    and updates Sources/index.md, the top-level index.md (Recent activity
+    + Counts), and log.md. Per SKILL.md rule 7.
+
+    Args:
+        content: Full text body to capture (markdown / plain text). For
+            files or binaries, use the /upload endpoint instead.
+        source_type: One of article, session, book, paper, video, other.
+        title: Human title; used to derive the filename slug.
+        url: Required when source_type is article, paper, or video.
+        tags: Lowercase dash-separated; the server normalizes case/spacing.
+        why_captured: One short paragraph on why this is worth keeping.
+            Rendered as a leading blockquote in the source body, between
+            the `# Source: ...` header and the `## Capture` section.
+
+    Returns:
+        {path, warnings}. On schema violation, raises a structured error
+        with code=INVALID_SOURCE, the missing fields, and the reason.
+    """
+    try:
+        result = add_module.add(
+            vault_root,
+            source_schema,
+            content=content,
+            source_type=source_type,
+            title=title,
+            url=url,
+            tags=tags,
+            why_captured=why_captured,
+        )
+    except add_module.AddError as e:
+        # FastMCP serializes raised exceptions; we want a structured shape.
+        raise ValueError(
+            f"{e.code}: {e.reason} (missing: {e.missing})"
+        ) from e
+    query_log.log_write_call(tool="add", written_path=result.path, cited_sources=[])
+    return result.as_dict()
+
+
+def op_audit(
+    vault_root: Path,categories: list[str] | None = None) -> dict:
+    """Audit / lint / health-check the Knowledge Base: find orphans, broken wikilinks, supersession gaps, stale unprocessed sources, and stale-review candidates. Read-only.
+
+    Returns a structured report Claude can read to propose follow-up
+    edits via `note`/`add`. Does NOT modify anything.
+
+    Categories (default: all):
+    - `broken_wikilink`: `[[X]]` whose target file doesn't exist.
+      Skips wikilinks inside fenced code blocks and inline code spans.
+      Bare names resolve against filename stems AND frontmatter `title:`
+      (so date-prefixed sources with a title match are not flagged).
+    - `orphan_entity`: `Entities/...` file with no inbound wikilinks
+    - `unprocessed_source`: source with empty `ingested_into:` (no notes
+      have compiled from it yet)
+    - `index_drift`: top-level `index.md` Counts disagree with on-disk counts
+    - `tag_inconsistency`: case/separator variants of the same tag
+      (`warning_letter_incident` vs `warning-letter-incident` vs
+      `Warning-Letter-Incident`). Mechanical drift only; semantic
+      near-duplicates like `metabolism` vs `metabolic` aren't flagged.
+    - `frontmatter_compliance`: per-page-type required-field gaps,
+      a `tenant:` set without the expected project, patterns using singular
+      `project:` instead of plural `projects:`.
+    - `unregistered_project_key`: a `project`/`projects` value not in the
+      registry (typo or genuinely new scope).
+    - `embedding_drift`: vector sidecar rows out of sync with disk (a file
+      changed/added/removed since it was last embedded).
+    - `relevance_pairs_pending`: real-usage (query -> cited_path) labels not
+      yet in the golden retrieval set.
+    - `stale_review`: active compiled conclusion that is old AND rarely
+      surfaced in `find` AND low inbound-link degree — a measurement-only
+      review candidate (still true? keep / supersede / archive). Never
+      decays or down-ranks; `find` ordering is unchanged.
+    - `corpus_contradictions`: corpus-wide pairs of active read-write
+      compiled conclusions whose embeddings sit just below the near-dup
+      threshold (close enough to restate/refine/contradict). A proximity
+      measurement surfaced for review (reconcile or supersede); never
+      auto-acted. No-ops when embeddings are disabled.
+
+    Args:
+        categories: Optional filter; only run these checks. Each must be
+            one of the categories above. Omit to run all.
+
+    Returns:
+        {findings: [{category, severity, path, detail, proposed_fix}],
+         summary: {category: count}}.
+    """
+    report = audit_module.audit(vault_root, categories=categories)
+    return report.as_dict()
+
+
+def op_audit_fix(
+    vault_root: Path,dry_run: bool = False, rebuild_embeddings: bool = False) -> dict:
+    """Run audit + auto-apply safe fixes; propose-only for risky categories.
+
+    Closes the lint-finds-but-doesn't-fix loop. Safe categories get
+    rewritten in-place via atomic batch writes; risky categories
+    surface in `proposed` for human/LLM review.
+
+    Safe categories (auto-applied):
+    - Canonical wikilink form across all compiled material (body +
+      frontmatter). Skips Sources/ and Evidence/ (append-only).
+    - Frontmatter required-field backfill with safe defaults:
+      - production-log missing created/updated → use started/shipped/today
+      - research-note/insight/failure/pattern missing status → "active"
+      - research-note/insight/failure/pattern missing updated →
+        use created, else today
+      - experiment missing duration → computed from started+concluded
+      - source missing captured → use created (if present)
+    - Pattern with singular `project:` → plural `projects: [<value>]`
+      (auto-merged into existing projects: list if present).
+    - Sub-folder index refresh + top-index count refresh.
+
+    Risky categories (propose-only, surfaced in `proposed` list):
+    - broken_wikilink residuals after canonicalization (forward refs,
+      missing files, audit limitations).
+    - orphan_entity (deletion is too big to auto-apply).
+    - unprocessed_source (compilation is a thinking task).
+    - tag_inconsistency (renames can break user mental models).
+    - frontmatter_compliance: tenant: misuse (might be intentional).
+    - source missing source_type (folder→type inference is brittle).
+
+    Idempotent: running twice on a clean vault produces no changes.
+
+    Args:
+        dry_run: If true, compute what would change without writing.
+            Default false.
+        rebuild_embeddings: If true, wipe and rebuild the vector sidecar
+            at `<vault>/Knowledge Base/.embeddings.sqlite` after the fix
+            sweep. Use on first run, after a machine swap, or when the
+            sidecar has drifted from disk. Ignored when `dry_run=true`.
+
+    Returns:
+        {fixed: [{category, path, detail, action}, ...],
+         proposed: [<audit findings>],
+         files_rewritten: int,
+         summary: {fixed: N, proposed: N, fixed_<category>: N,
+                   embeddings_chunks?: N},
+         dry_run: bool}
+    """
+    report = audit_fix_module.audit_fix(
+        vault_root,
+        dry_run=dry_run,
+        rebuild_embeddings=rebuild_embeddings,
+    )
+    return report.as_dict()
+
+
+def op_reconcile(
+    vault_root: Path,dry_run: bool = False) -> dict:
+    """Heal vault drift from out-of-band edits in one pass.
+
+    The writers keep the embedding sidecar, index.md count rows, and log.md
+    current on every write. But editing the vault directly — in Obsidian,
+    on mobile, or via a manual filesystem edit — bypasses those hooks, so
+    the sidecar and the counts drift silently. `reconcile` is the
+    first-class "I edited around the system, fix it" command:
+
+    1. Index counts — recompute Sources/Notes/Entities count rows from
+       on-disk reality and rewrite any that drifted (curated descriptions
+       and Recent-activity are preserved; only count tokens move).
+    2. Embeddings — incrementally re-embed only the *stale* files (those
+       `embedding_drift` flags: on-disk mtime newer than the sidecar row),
+       via the same path the writers use. Cheaper than
+       `audit_fix(rebuild_embeddings=true)`'s full wipe-and-rebuild.
+    3. Drift report — re-run index_drift + embedding_drift, return what
+       remains.
+
+    Narrower than `audit_fix`: it does NOT canonicalize wikilinks or
+    backfill frontmatter (those are content rewrites you opt into).
+    Idempotent; `dry_run=true` reports without writing.
+
+    Args:
+        dry_run: If true, compute what would change without writing.
+            Default false.
+
+    Returns:
+        {indexes_updated: [<index path>, ...],
+         embeddings_refreshed: int,
+         embeddings_status: "current" | "refreshed" | "disabled",
+         remaining_drift: [<audit findings>],
+         dry_run: bool}
+    """
+    report = reconcile_module.reconcile(vault_root, dry_run=dry_run)
+    return report.as_dict()
+
+
+def op_provenance_report(
+    vault_root: Path,
+    tag: str | None = None,
+    key: str | None = None,
+    value: str | None = None,
+    path: str | None = None,
+) -> dict:
+    """Trace provenance: scan note bodies for `<!-- key:value -->` tags — where an opinion/take/flag came from. Read-only.
+
+    On-demand scan over markdown bodies — no index, no sidecar. Use it to
+    answer "show all conv:-derived takes" or "what's flagged add-to-imdb"
+    without grepping. The opinion/taste rows carry provenance as HTML
+    comments (e.g. `<!-- platform:imdb -->`, `<!-- conv:2026-06-01 -->`);
+    this reads them in place. Tags inside fenced code are ignored; multiple
+    comments and multiple key:value pairs on one line are all parsed.
+
+    Args:
+        tag: Shorthand filter — "key" or "key:value" (e.g. "platform:imdb").
+        key: Filter to rows carrying this provenance key.
+        value: With key, require this exact value.
+        path: Restrict the scan to one vault-relative file (else the whole
+            Knowledge Base is walked).
+
+    Returns:
+        {findings: [{path, line_number, row_text, tags}], summary:
+         {key: count}}. line_number is body-relative (frontmatter excluded).
+    """
+    findings = provenance_module.scan_provenance(
+        vault_root, tag=tag, key=key, value=value, path=path
+    )
+    summary: dict[str, int] = {}
+    for f in findings:
+        for k in f.tags:
+            summary[k] = summary.get(k, 0) + 1
+    return {"findings": [f.as_dict() for f in findings], "summary": summary}
+
+
+def op_propose_compilation(
+    vault_root: Path,
+    sources: list[str],
+    suggested_title: str | None = None,
+) -> dict:
+    """Draft / scaffold a compiled note from unprocessed source(s) — what to compile next, drain the source backlog. Read-only.
+
+    The backlog-drain companion to `audit`'s `unprocessed_source` findings:
+    point it at one or more raw sources and it hands back a ready-to-fill
+    note skeleton — inferred note_type, a Question/Findings/Connections (or
+    Claim/…) outline, the `sources[]` to cite, and adjacent compiled pages to
+    link (computed via the same hybrid retrieval as `suggest_links`). It
+    does NOT write anything: you fill the prose and call `note()` with the
+    returned `suggested_sources` + `suggested_connections`.
+
+    Group sources yourself before calling — pass a set that genuinely belongs
+    in one note (the audit list is aged oldest-first to help you triage).
+
+    Args:
+        sources: Vault-relative paths/wikilinks to the source(s) to compile.
+            Same path conventions as `note.sources` (brackets and the
+            leading `Knowledge Base/` are tolerated).
+        suggested_title: Optional title override; otherwise one is derived
+            from the source titles.
+
+    Returns:
+        {suggested_note_type, suggested_title, suggested_sources,
+         suggested_connections, outline_markdown, warnings}.
+
+    Errors:
+        INVALID_PROPOSE (no sources); SOURCES_NOT_FOUND (none resolved).
+    """
+    try:
+        return compile_proposal_module.propose_compilation(
+            vault_root, sources=sources, suggested_title=suggested_title
+        )
+    except compile_proposal_module.ProposeError as e:
+        raise ValueError(f"{e.code}: {e.reason}") from e
+
+
+def op_get(
+    vault_root: Path,
+    path: str,
+    frontmatter_only: bool = False,
+    include_history: bool = False,
+    links: bool = False,
+) -> dict:
+    """Read / open / fetch / load the full contents of a KB or vault page by path. Returns frontmatter + body + raw content.
+
+    Reads anywhere under the vault root — not just `Knowledge Base/`.
+    This lets you cite from curated, read-only sibling folders (e.g.
+    `Reference/`) kept outside Knowledge Base/ when compiling. Those are
+    read-only by convention (marked in `_access.yaml`); `get` honors that
+    by only reading.
+
+    Use this when `find` gives you a path and you need the whole page
+    (to cite, build on, or rewrite). `find` only returns excerpts.
+
+    Args:
+        path: Vault-relative path. Accepted shapes:
+            - `Knowledge Base/Notes/Insights/foo.md`
+            - `Reference/Strategy.md`
+            - `Notes/Insights/foo` (auto-prepends `Knowledge Base/` if
+              literal path doesn't resolve; auto-adds `.md`).
+        frontmatter_only: If true, return ONLY the frontmatter (no body) —
+            cheap for scanning many files by field (folds in the former
+            `get_frontmatter` tool). Returns {path, frontmatter,
+            has_frontmatter} instead of the full page below.
+        include_history: If true, attach a `history` list — the page's
+            change log from the append-only `log.md`, newest-first
+            (`[{date, op, summary}]`, where `summary` is the `why:`
+            rationale recorded at write time). Use this to answer "why was
+            this note changed / what was the old version / show its history"
+            and to verify an edit's rationale. `[]` when the page has no
+            recorded edits.
+        links: If true, attach a `links` summary —
+            `{inbound: [...], outbound: [...]}`. `inbound` lists files whose
+            wikilinks resolve to this page (each
+            `{path, line_number, context, raw_target}`); `outbound` lists
+            the distinct wikilink targets in this page's body. Use it to
+            see a note's graph neighbourhood in one call. Default off (no
+            behaviour change).
+
+    Returns:
+        {path, frontmatter, body, content, content_hash, mtime}.
+        `content` is the raw file text (including frontmatter delimiters);
+        `body` is just the markdown after the frontmatter. `content_hash`
+        is a sha256 you can echo back to `edit`/`multi_edit` via
+        `expected_hash` to refuse a write if the file changed on disk since
+        this read (two-writer drift guard); `mtime` is advisory.
+        Adds `history` when `include_history=true`.
+
+    Errors:
+        INVALID_PATH (path escapes vault root or empty);
+        NOT_FOUND (no such file); UNREADABLE (parse failure).
+    """
+    if frontmatter_only:
+        try:
+            fm_result = get_frontmatter_module.get_frontmatter(
+                vault_root, path=path
+            )
+        except get_frontmatter_module.GetFrontmatterError as e:
+            raise ValueError(f"{e.code}: {e.reason}") from e
+        out = fm_result.as_dict()
+    else:
+        try:
+            result = get_page_module.get_page(vault_root, path=path)
+        except get_page_module.GetError as e:
+            raise ValueError(f"{e.code}: {e.reason}") from e
+        out = result.as_dict()
+    query_log.log_get_call(
+        read_path=out["path"],
+        frontmatter_only=frontmatter_only,
+        include_history=include_history,
+    )
+    if include_history:
+        out["history"] = vault.read_log_entries(vault_root, out["path"])
+    if links:
+        out["links"] = _link_summary(
+            vault_root, out.get("path", ""), out.get("body", "")
+        )
+    return out
+
+
+def op_edit(
+    vault_root: Path,
+    path: str,
+    why: str,
+    new_body: str | None = None,
+    tags: list[str] | None = None,
+    old_string: str | None = None,
+    new_string: str | None = None,
+    replace_all: bool = False,
+    heading: str | None = None,
+    section_position: str = "append",
+    edits: list[dict] | None = None,
+    row_key: str | None = None,
+    take: str | None = None,
+    overwrite: bool = False,
+    field: str | None = None,
+    value: str | int | float | bool | list | dict | None = None,
+    allow_curated: bool = False,
+    expected_hash: str | None = None,
+    validate_only: bool = False,
+) -> dict:
+    """Lightweight in-place edit of a page (body, tags, a surgical snippet,
+    a batch, an opinion row, or one frontmatter field).
+
+    For tweaks — typo fixes, filling a row, appending one line, tag
+    corrections — without going through full supersession via `replace`.
+    Use `replace` for substantial rewrites; use `edit` when creating a new
+    file + superseded-link chain would be silly for what you're changing.
+
+    One mode per call. Three param-selected modes fold in former tools:
+    - `edits=[...]` -> batch surgical edits in one atomic commit (was the
+      `multi_edit` tool). Each item {old_string, new_string, replace_all?}
+      applies sequentially.
+    - `row_key=...` + `take=...` -> fill a `[take: ]` opinion row by its
+      leading text without re-sending the body (was `set_take`).
+    - `field=...` + `value=...` -> patch ONE frontmatter field; pass
+      `allow_curated=true` for curated trees (was `set_frontmatter_field`).
+    Otherwise the default (composable) body/tags/surgical modes:
+    - `new_body` — replace the WHOLE body. Heavyweight; you re-send
+      everything after the frontmatter.
+    - `tags` — replace the `tags:` frontmatter field.
+    - `old_string`/`new_string` — **surgical** string-replace inside the
+      body. Token-cheap: send only the changed snippet, not the whole
+      page. Ideal for filling a `[take: ]` row or appending one opinion
+      (replace a section heading with itself + the new line). `updated:`
+      is always bumped to today.
+
+    Surgical-mode rules (mirrors a precise find-and-replace):
+    - `old_string` must match the file EXACTLY, including whitespace.
+    - By default it must occur exactly once — an ambiguous match is an
+      error (AMBIGUOUS_MATCH) so you never edit the wrong row. Pass
+      `replace_all=True` to replace every occurrence.
+    - Cannot be combined with `new_body` (both rewrite the body); may be
+      paired with `tags`.
+    - Only the inserted snippet gets wikilink-normalized; the rest of the
+      body is left byte-for-byte untouched.
+
+    What stays in all modes:
+    - All other frontmatter fields (type, project, status, sources,
+      superseded_by, etc.). If you need to change those, use `replace`.
+
+    No type allowlist: any frontmatter-bearing page outside Sources/
+    Evidence is editable, regardless of `type:`. Works on novel page
+    types (`identity`, future types) without code changes.
+
+    Refuses:
+    - Sources/ and Evidence/ paths (rule 2: append-only). Add a
+      corrective source or compile a downstream note instead.
+    - Pages without a frontmatter block (won't synthesize one).
+    - Pages already marked `status: superseded` (don't edit history;
+      supersede the active page instead).
+
+    Args:
+        path: Vault-relative path to the compiled page (same shape as
+            `get` accepts).
+        why: One-line rationale for the edit. Required — lands in the
+            log entry so the change is auditable.
+        new_body: New markdown body (everything after frontmatter).
+            Omit to keep the existing body.
+        tags: New tags list (replaces existing). Lowercase dash-
+            separated; the server normalizes. Omit to keep existing tags.
+        old_string: Exact snippet to find in the body (surgical mode).
+        new_string: Replacement snippet (required with old_string; must
+            differ from it).
+        replace_all: Replace every occurrence instead of requiring a
+            unique match. Default False.
+        heading: Section-targeted mode — the `## Heading` (the `#` markers
+            are optional) under which to place `new_string`. The section
+            spans from that heading to the next heading of equal-or-higher
+            level (or EOF). Mutually exclusive with new_body/old_string.
+            Raises HEADING_NOT_FOUND if absent.
+        section_position: With `heading`, where to put `new_string`:
+            "append" (default), "prepend", or "replace" the section body.
+        edits: Batch-surgical mode — list of {old_string, new_string,
+            replace_all?} applied sequentially in one atomic commit.
+        row_key: Take-row mode — natural leading text of the row to fill
+            (e.g. "Whiplash (2014)"). Requires `take`.
+        take: Text to write between `[take:` and `]` (take-row mode).
+        overwrite: In take-row mode, also replace an already-filled take.
+        field: Frontmatter-patch mode — the single frontmatter key to set
+            (cannot be `updated`, which is auto-bumped).
+        value: New value for `field` (scalar/list/dict).
+        allow_curated: Allow a frontmatter patch under a curated tree.
+        expected_hash: Optional drift guard. Pass the `content_hash` you
+            got from `get`; the edit refuses (STALE_EDIT) if the file
+            changed on disk since, so you never clobber another writer.
+        validate_only: Preview a surgical match without writing. Needs
+            `old_string`. Reports how many rows would be hit instead of
+            committing — use it before a `replace_all` to avoid an
+            ambiguous match silently touching more rows than intended.
+
+    Returns:
+        Shape varies by mode (take-row -> {path, row, warnings};
+        frontmatter-patch -> {path, field, old_value, new_value, warnings};
+        batch -> {path, edits_applied, warnings}). Default mode normally
+        {path, warnings}. When validate_only=True:
+        {path, validate_only, mode, match_count, matches} — `matches` is
+        the line(s) around each occurrence; nothing is written.
+
+    Errors:
+        INVALID_EDIT (nothing to edit, old_string+new_body both given,
+        new_string missing/equal, path in Sources/Evidence); NOT_FOUND;
+        STRING_NOT_FOUND (surgical snippet absent); AMBIGUOUS_MATCH
+        (snippet not unique and replace_all=False); ALREADY_SUPERSEDED;
+        STALE_EDIT (expected_hash mismatch — file changed since read);
+        UNREADABLE.
+    """
+    active = [n for n, on in (
+        ("edits", edits is not None),
+        ("row_key", row_key is not None),
+        ("field", field is not None),
+    ) if on]
+    if len(active) > 1:
+        raise ValueError(
+            f"INVALID_EDIT: one edit mode at a time; got {', '.join(active)}"
+        )
+    try:
+        if edits is not None:
+            result = multi_edit_module.multi_edit(
+                vault_root, path=path, why=why, edits=edits,
+                expected_hash=expected_hash, validate_only=validate_only,
+            )
+        elif row_key is not None:
+            if take is None:
+                raise ValueError("INVALID_EDIT: row_key mode requires `take`")
+            result = set_take_module.set_take(
+                vault_root, path=path, row_key=row_key, take=take,
+                why=why, overwrite=overwrite,
+            )
+        elif field is not None:
+            result = set_frontmatter_field_module.set_frontmatter_field(
+                vault_root, path=path, field=field, value=value,
+                why=why, allow_curated=allow_curated,
+            )
+        else:
+            result = edit_module.edit(
+                vault_root, path=path, why=why, new_body=new_body,
+                tags=tags, old_string=old_string, new_string=new_string,
+                replace_all=replace_all, heading=heading,
+                section_position=section_position,
+                expected_hash=expected_hash, validate_only=validate_only,
+            )
+    except (
+        edit_module.EditError,
+        set_take_module.SetTakeError,
+        set_frontmatter_field_module.SetFrontmatterError,
+    ) as e:
+        msg = f"{e.code}: {e.reason}"
+        if getattr(e, "missing", None):
+            msg += f" (missing: {e.missing})"
+        if getattr(e, "candidates", None):
+            msg += f" (candidates: {e.candidates})"
+        raise ValueError(msg) from e
+    return result.as_dict()
+
+
+def op_replace(
+    vault_root: Path,
+    old_path: str,
+    content: str,
+    note_type: str,
+    title: str,
+    reason: str | None = None,
+    project: str | None = None,
+    projects: list[str] | None = None,
+    sources: list[str] | None = None,
+    tags: list[str] | None = None,
+    status: str | None = None,
+    severity: str | None = None,
+    pattern_type: str | None = None,
+    domain: str | None = None,
+    started: str | None = None,
+    duration: str | None = None,
+    hypothesis: str | None = None,
+    n: int | None = None,
+    concluded: str | None = None,
+    medium: str | None = None,
+    recorded: str | None = None,
+    published: str | None = None,
+    host: str | None = None,
+    editor: str | None = None,
+    project_category: str | None = None,
+) -> dict:
+    """Supersede an existing compiled page with a new one.
+
+    Writes the new page at a fresh slug (via the same machinery as
+    `note`), then patches the OLD page to set `status: superseded` and
+    `superseded_by: "[[<new>]]"`. The NEW page gets `supersedes:
+    "[[<old>]]"` in its frontmatter. The old page stays readable;
+    readers follow the chain — inbound wikilinks are NOT retargeted
+    (per SKILL.md rule 6).
+
+    Use this for substantial rewrites of an existing page — not minor
+    tweaks (the desk-side flow handles those better since you see a
+    live diff). Cannot supersede sources or evidence (append-only).
+    No type allowlist beyond the append-only guard: novel page types
+    (`identity`, future types) can be superseded without code changes.
+
+    Args:
+        old_path: Vault-relative path of the page being superseded.
+            Same path conventions as `get` and `find`.
+        reason: Optional one-line explanation of why this replacement is
+            happening; lands in the log entry body.
+        (all other args): Same as the `note` tool — define the new page's
+            content, type, project/projects, sources, etc.
+
+    Returns:
+        {old_path, new_path, warnings}.
+
+    Errors:
+        INVALID_REPLACE (old is in Sources/ or Evidence/, or not a
+        supersedable type); OLD_NOT_FOUND; ALREADY_SUPERSEDED
+        (old page is already marked superseded).
+    """
+    try:
+        result = replace_module.replace(
+            vault_root,
+            old_path=old_path,
+            reason=reason,
+            content=content,
+            note_type=note_type,
+            title=title,
+            project=project,
+            projects=projects,
+            sources=sources,
+            tags=tags,
+            status=status,
+            severity=severity,
+            pattern_type=pattern_type,
+            domain=domain,
+            started=started,
+            duration=duration,
+            hypothesis=hypothesis,
+            n=n,
+            concluded=concluded,
+            medium=medium,
+            recorded=recorded,
+            published=published,
+            host=host,
+            editor=editor,
+            project_category=project_category,
+        )
+    except replace_module.ReplaceError as e:
+        raise ValueError(
+            f"{e.code}: {e.reason} (missing: {e.missing})"
+        ) from e
+    except note_module.NoteError as e:
+        # New-page validation failed before the supersession could land.
+        raise ValueError(
+            f"{e.code}: {e.reason} (missing: {e.missing})"
+        ) from e
+    query_log.log_write_call(
+        tool="replace", written_path=result.new_path, cited_sources=sources
+    )
+    return result.as_dict()
+
+
+def op_link(
+    vault_root: Path,
+    entity_type: str,
+    name: str,
+    summary: str,
+    why_in_kb: str | None = None,
+    tags: list[str] | None = None,
+    connections: list[str] | None = None,
+    affiliation: str | None = None,
+    relationship: str | None = None,
+    domain: str | None = None,
+    language: str | None = None,
+    repo: str | None = None,
+    license: str | None = None,
+    used_in: list[str] | None = None,
+    decided: str | None = None,
+    project: str | None = None,
+    decision_status: str | None = None,
+) -> dict:
+    """Create a typed entity under Entities/<Folder>/<Name>.md.
+
+    Entities are the typed nodes of the KB graph — people, concepts,
+    libraries, decisions. Name them after the thing they are (Title Case,
+    not slugified): `Andrej Karpathy`, `Agentic RAG`, `pgvector`.
+
+    Four entity types with conditional frontmatter:
+    - `person`   → Entities/People/. Optional: `affiliation`, `relationship`.
+    - `concept`  → Entities/Concepts/. Optional: `domain` (e.g.
+      "retrieval", "metabolism", "infrastructure").
+    - `library`  → Entities/Libraries/. Optional: `language`, `repo`,
+      `license`, `used_in` (list of projects).
+    - `decision` → Entities/Decisions/. Optional: `decided` (YYYY-MM-DD),
+      `project` (project key — any slug; unknown keys auto-register on first
+      use, same as `note`), `decision_status` ∈ {proposed, accepted,
+      superseded}.
+
+    v1 is create-only. If the entity file already exists, returns
+    ENTITY_EXISTS — use `replace` to supersede instead. Sub-folder index
+    (e.g. Entities/Concepts/index.md categorization) is NOT auto-updated;
+    reconcile via desk audit.
+
+    Args:
+        entity_type: One of person, concept, library, decision.
+        name: Title Case, the entity's actual name. Will be the filename.
+        summary: One-paragraph description for the `## Summary` section.
+        why_in_kb: Optional `## Why in the KB` paragraph — explains what
+            this entity is relevant to in your work.
+        tags: Lowercase dash-separated; normalized by the server.
+        connections: List of vault-relative wikilink targets to put under
+            `## Connections`. Same path conventions as `note.sources`.
+        (per-type fields): see the bullet list above.
+
+    Returns:
+        {path, warnings}.
+
+    Errors:
+        INVALID_LINK (bad entity_type, decision_status, missing required);
+        ENTITY_EXISTS (use `replace` instead).
+    """
+    try:
+        result = link_module.link(
+            vault_root,
+            entity_type=entity_type,
+            name=name,
+            summary=summary,
+            why_in_kb=why_in_kb,
+            tags=tags,
+            connections=connections,
+            affiliation=affiliation,
+            relationship=relationship,
+            domain=domain,
+            language=language,
+            repo=repo,
+            license=license,
+            used_in=used_in,
+            decided=decided,
+            project=project,
+            decision_status=decision_status,
+        )
+    except link_module.LinkError as e:
+        raise ValueError(
+            f"{e.code}: {e.reason} (missing: {e.missing})"
+        ) from e
+    return result.as_dict()
+
+
+def op_preserve(
+    vault_root: Path,
+    scope: str,
+    category: str,
+    filename: str,
+    content: str,
+    description: str | None = None,
+) -> dict:
+    """Capture a TEXT artifact to Evidence/<scope>/<category>/.
+
+    For raw factual artifacts that are text — transcripts, pasted letters,
+    email bodies — preserved as-received with no analytical processing. Per
+    SKILL.md rule 2, Evidence is append-only; analytical takes go in compiled
+    notes that link to the evidence file.
+
+    BINARY artifacts (PDFs, images, .docx — any non-text file) are delivered
+    out-of-band, not through this tool: call `mint_upload_token` and POST the
+    bytes to `/upload`, or drop the file into Evidence/ desk-side via Obsidian
+    Sync. The bytes never pass through the model.
+
+    Args:
+        scope: Incident or domain key (e.g. "Yolo", "Mother Cancer").
+            Creates the subfolder if it doesn't exist.
+        category: Sub-category within scope (e.g. "letters", "labs",
+            "court-docs"). Creates the subfolder if it doesn't exist.
+        filename: The artifact's filename including extension
+            (e.g. `2026-04-15-statement.txt`).
+        content: UTF-8 text to preserve as-received.
+        description: Optional. If supplied, a sidecar `<filename>.md`
+            is written alongside the artifact with frontmatter and the
+            description under `## Description`.
+
+    Returns:
+        {path, sidecar_path, warnings}.
+
+    Errors:
+        INVALID_PRESERVE (missing required); ARTIFACT_EXISTS (file already
+        exists — Evidence is append-only, pick a new filename).
+    """
+    try:
+        result = preserve_module.preserve(
+            vault_root,
+            scope=scope,
+            category=category,
+            filename=filename,
+            content=content,
+            description=description,
+        )
+    except preserve_module.PreserveError as e:
+        raise ValueError(
+            f"{e.code}: {e.reason} (missing: {e.missing})"
+        ) from e
+    return result.as_dict()
+
+
+def op_note(
+    vault_root: Path,
+    content: str,
+    note_type: str,
+    title: str,
+    project: str | None = None,
+    projects: list[str] | None = None,
+    sources: list[str] | None = None,
+    tags: list[str] | None = None,
+    status: str | None = None,
+    severity: str | None = None,
+    pattern_type: str | None = None,
+    domain: str | None = None,
+    started: str | None = None,
+    duration: str | None = None,
+    hypothesis: str | None = None,
+    n: int | None = None,
+    concluded: str | None = None,
+    medium: str | None = None,
+    recorded: str | None = None,
+    published: str | None = None,
+    host: str | None = None,
+    editor: str | None = None,
+    project_category: str | None = None,
+) -> dict:
+    """Create a compiled note in the Knowledge Base.
+
+    Use this for distilled thinking — not raw capture. For raw capture
+    (an article you read, a session transcript), use `add` instead.
+
+    Six note types:
+    - `research-note`: project-scoped findings. `project` REQUIRED.
+      → `Notes/Research/<Project>/<slug>.md`
+    - `insight`: cross-cutting claim. Optional `projects` (plural).
+      → `Notes/Insights/<slug>.md`
+    - `failure`: documented failure mode. Optional `projects`, optional
+      `severity` ∈ {minor, moderate, serious, critical}.
+      → `Notes/Failures/<slug>.md`
+    - `pattern`: reusable cross-cutting pattern. Optional `projects`,
+      optional `pattern_type` ∈ {architectural, workflow, prompting,
+      governance, pedagogical}.
+      → `Notes/Patterns/<slug>.md`
+    - `experiment`: hypothesis + protocol. `domain`, `started` (YYYY-MM-DD),
+      and `duration` (e.g. "30 days", "ongoing") REQUIRED. Optional
+      `hypothesis`, `n` (default 1), `concluded`.
+      → `Notes/Experiments/<domain>/YYYY-MM-<slug>.md`
+    - `production-log`: creative artifact log. `medium` REQUIRED (e.g.
+      "Reels", "Episodes"). Optional `recorded`, `published`, `host`,
+      `editor`, `projects`. Status enum is richer: {planned, recorded,
+      edited, published, reflected, dropped, archived}; defaults to
+      `planned`.
+      → `Notes/Productions/<medium>/YYYY-MM-<slug>.md`
+
+    For each `sources:` wikilink, appends this note's wikilink to that
+    source's `ingested_into:` frontmatter (maintaining the source→note graph).
+
+    Args:
+        content: Full markdown body, written verbatim after the
+            frontmatter. Should start with `# <title>` (the H1 matching
+            the title arg) followed by the section conventions per type:
+            research-note: `## Question`/`## Findings`/`## Connections`.
+            insight: `## Claim`/`## Why it holds`/`## Connections`.
+            failure: `## What happened`/`## Mechanism`/`## Detection`/`## Mitigation`/`## Connections`.
+            pattern: `## Problem`/`## Solution`/`## When to use`/`## When NOT to use`/`## Connections`.
+            experiment: `## Hypothesis`/`## Protocol`/`## Baseline`/`## Intervention`/`## Results`/`## Conclusion`/`## Connections`.
+            production-log: `## Frame`/`## Artifact`/`## Production session`/`## Outcomes`/`## Reflection`/`## Connections`.
+            Conventions only — no shape is enforced.
+        note_type: One of research-note, insight, failure, pattern,
+            experiment, production-log.
+        title: Human title; used to derive a kebab-case filename slug.
+            Experiments and production-logs auto-prefix with YYYY-MM.
+        project: REQUIRED for research-note. __PROJECT_KEYS_HINT__
+        projects: List of project keys (plural). Optional for insight,
+            failure, pattern, production-log. __PROJECT_KEYS_HINT__
+        sources: Vault-relative wikilinks to existing pages this note draws
+            from, e.g. `["Knowledge Base/Sources/Articles/2026-05-18-foo"]`
+            or `["[[Knowledge Base/Sources/Articles/2026-05-18-foo]]"]`.
+            Brackets and the leading `Knowledge Base/` are tolerated.
+        tags: Lowercase dash-separated; the server normalizes case/spacing.
+        status: Defaults to `active` for most types, `planned` for
+            production-log. Valid set varies by type.
+        severity: failure only. {minor, moderate, serious, critical}.
+        pattern_type: pattern only. {architectural, workflow, prompting,
+            governance, pedagogical}.
+        domain: experiment only. Becomes the subfolder name (lowercased).
+        started: experiment only. YYYY-MM-DD when the experiment began.
+        duration: experiment only. Freeform, e.g. "30 days", "ongoing".
+        hypothesis: experiment only. One-line claim being tested.
+        n: experiment only. Sample size. Defaults to 1 (n-of-1).
+        concluded: experiment only. YYYY-MM-DD when it ended (absent while ongoing).
+        medium: production-log only. Subfolder, e.g. "Reels", "Episodes".
+        recorded: production-log only. YYYY-MM-DD of recording session.
+        published: production-log only. YYYY-MM-DD of publication.
+        host: production-log only. Creator/talent name.
+        editor: production-log only. Producer/editor name.
+
+    Returns:
+        {path, warnings}. On validation failure, raises a structured error
+        with code=INVALID_NOTE, the missing fields, and the reason.
+    """
+    try:
+        result = note_module.note(
+            vault_root,
+            content=content,
+            note_type=note_type,
+            title=title,
+            project=project,
+            projects=projects,
+            sources=sources,
+            tags=tags,
+            status=status,
+            severity=severity,
+            pattern_type=pattern_type,
+            domain=domain,
+            started=started,
+            duration=duration,
+            hypothesis=hypothesis,
+            n=n,
+            concluded=concluded,
+            medium=medium,
+            recorded=recorded,
+            published=published,
+            host=host,
+            editor=editor,
+            project_category=project_category,
+        )
+    except note_module.NoteError as e:
+        raise ValueError(
+            f"{e.code}: {e.reason} (missing: {e.missing})"
+        ) from e
+    query_log.log_write_call(
+        tool="note", written_path=result.path, cited_sources=sources
+    )
+    return result.as_dict()
+
+
+def op_query_data(
+    vault_root: Path,
+    path: str,
+    record_path: str | None = None,
+    filters: list[dict] | None = None,
+    columns: list[str] | None = None,
+    sort_by: str | None = None,
+    descending: bool = False,
+    limit: int = 100,
+    offset: int = 0,
+    aggregate: str | None = None,
+    date_from: str | None = None,
+    date_to: str | None = None,
+    date_column: str | None = None,
+) -> dict:
+    """Tier 2: structured query over a CSV/JSON data file under the vault.
+
+    The retrieval half of the data-search pattern — `find` surfaces a
+    dataset's markdown "card"; this reads the raw file the card points at
+    and returns exact rows / aggregates (no whole-file dump). KB datasets
+    are small, so it reads on demand — no index, no new infra.
+
+    Formats: CSV / TSV, and JSON (a top-level array, or a nested array via
+    `record_path` / common-key auto-detect). Column names may be dotted to
+    reach nested JSON fields (e.g. "performer.name", "id.extension")
+    anywhere a column is named (filters / columns / sort / aggregate).
+
+    Args:
+        path: vault-relative path to the `.csv` / `.tsv` / `.json` file.
+        record_path: (JSON) dotted path to the array inside a nested
+            object, e.g. "sections.work_incapacity". Omit for a top-level
+            array or the common keys result/results/data/rows/items/entries.
+        filters: list of `{column, op, value}`. `op` ∈ eq, ne, gt, gte, lt,
+            lte, contains, icontains, startswith, in, nin, exists, missing.
+            Numeric compares coerce tolerantly (comma decimals; lab
+            operators like "<0.4"/">75" are stripped for the comparison).
+        columns: project to these columns (dotted ok). Omit for all.
+        sort_by / descending: sort by a column (numeric-aware).
+        limit / offset: pagination (limit default 100, hard cap 1000).
+        aggregate: instead of rows — "count"; "func:column" where func ∈
+            min, max, sum, avg, latest, distinct; or "profile" to get a
+            deterministic content profile (per-column kind, distinct values,
+            numeric ranges, date span) under `aggregate.profile` PLUS a
+            ready-to-write markdown dataset card under `aggregate.dataset_card`.
+            Use "profile" to make a CSV/JSON findable — write the card into
+            the KB (fill in its "What this holds" line) so the dataset is
+            discoverable by content without ever embedding its raw rows.
+        date_from / date_to / date_column: convenience date-range filter on
+            `date_column` (defaults to a "date" column if present); ISO
+            date strings, compared lexicographically.
+
+    Returns:
+        {path, format, total_rows, total_matched, returned, columns, rows,
+         aggregate, truncated, warnings}.
+
+    Errors: INVALID_PATH / NOT_FOUND (path); UNSUPPORTED_FORMAT; TOO_LARGE;
+        BAD_JSON; BAD_RECORD_PATH; BAD_FILTER; BAD_OP; BAD_AGGREGATE.
+    """
+    try:
+        result = query_data_module.query_data(
+            vault_root,
+            path=path,
+            record_path=record_path,
+            filters=filters,
+            columns=columns,
+            sort_by=sort_by,
+            descending=descending,
+            limit=limit,
+            offset=offset,
+            aggregate=aggregate,
+            date_from=date_from,
+            date_to=date_to,
+            date_column=date_column,
+        )
+    except query_data_module.QueryDataError as e:
+        raise ValueError(f"{e.code}: {e.reason}") from e
+    return result.as_dict()
+
+
+def op_create_file(
+    vault_root: Path,
+    path: str,
+    content: str = "",
+    frontmatter: dict | None = None,
+    overwrite: bool = False,
+    allow_curated: bool = False,
+    kind: str = "file",
+    parents: bool = True,
+) -> dict:
+    """Tier 2: write a file — or, with `kind="dir"`, create a folder — at an
+    arbitrary vault path.
+
+    With `kind="dir"`, this creates a folder (mkdir -p when `parents=true`)
+    and ignores `content`/`frontmatter`/`overwrite` (folds in the former
+    `create_directory` tool); returns {path, created, warnings}.
+
+    Escape hatch for files that don't fit Tier 1 type routing — new folder
+    structures (`Identity/`, `Templates/`), skill files, scratch. For
+    typed notes use `note`/`add`/`link`/`preserve`.
+
+    If `frontmatter` is a dict, this op prepends a YAML block built from
+    it (and auto-fills `created`/`updated` to today if not provided);
+    `content` is the body in that case. If `frontmatter` is omitted,
+    `content` is written verbatim — the caller is responsible for any
+    frontmatter already in it.
+
+    Refuses:
+    - Sources/, Evidence/ (append-only — use `add` or `preserve`).
+    - Subtrees marked `readonly`/`excluded` in `_access.yaml` (curated,
+      read-only material) — a hard refusal with no override.
+    - Existing files unless `overwrite=true`.
+
+    Args:
+        path: Vault-relative, e.g. `Knowledge Base/Identity/Career.md`.
+            Forward or back slashes accepted. Path-escape guarded.
+        content: File body (or full file if `frontmatter` is None). Text
+            only; for binaries use the /upload endpoint.
+        frontmatter: Optional dict prepended as YAML frontmatter.
+        overwrite: If true, replace existing file. Default false.
+        allow_curated: Required to write under a curated tree. Default false.
+        kind: "file" (default) or "dir". With "dir", creates a folder
+            instead of a file (former `create_directory`).
+        parents: In "dir" mode, create intermediate folders (mkdir -p).
+            Default true.
+
+    Returns: {path, warnings} for files; {path, created, warnings} for dirs.
+    Errors: INVALID_PATH; APPEND_ONLY; CURATED_PROTECTED; FILE_EXISTS;
+            NOT_A_FILE; (dir mode) NOT_A_DIR; MISSING_PARENT; MKDIR_FAILED.
+    """
+    if kind == "dir":
+        try:
+            result = create_directory_module.create_directory(
+                vault_root,
+                path=path,
+                parents=parents,
+                allow_curated=allow_curated,
+            )
+        except create_directory_module.CreateDirectoryError as e:
+            raise ValueError(f"{e.code}: {e.reason}") from e
+        return result.as_dict()
+    try:
+        result = create_file_module.create_file(
+            vault_root,
+            path=path,
+            content=content,
+            frontmatter=frontmatter,
+            overwrite=overwrite,
+            allow_curated=allow_curated,
+        )
+    except create_file_module.CreateFileError as e:
+        raise ValueError(f"{e.code}: {e.reason}") from e
+    return result.as_dict()
+
+
+def op_list_directory(
+    vault_root: Path,
+    path: str = "",
+    recursive: bool = False,
+    include_hidden: bool = False,
+) -> dict:
+    """Tier 2: list files and subfolders at a vault path. Read-only.
+
+    Works anywhere under vault root including curated trees (consistent
+    with `get`). For .md files, surfaces the frontmatter `type` field
+    so callers can scan typed content quickly.
+
+    Args:
+        path: Vault-relative. Empty string lists vault root. Auto-handles
+            forward/back slashes.
+        recursive: If true, walk subfolders. Default false.
+        include_hidden: If true, include dotfiles and _attachments/.
+            Default false.
+
+    Returns: {path, entries: [{name, type, path, size_bytes, updated,
+             frontmatter_type}]}.
+
+    Errors: INVALID_PATH; NOT_FOUND; NOT_A_DIR.
+    """
+    try:
+        result = list_directory_module.list_directory(
+            vault_root,
+            path=path,
+            recursive=recursive,
+            include_hidden=include_hidden,
+        )
+    except list_directory_module.ListDirectoryError as e:
+        raise ValueError(f"{e.code}: {e.reason}") from e
+    return result.as_dict()
+
+
+def op_move_file(
+    vault_root: Path,
+    old_path: str,
+    new_path: str,
+    update_wikilinks: bool = True,
+    allow_curated: bool = False,
+) -> dict:
+    """Tier 2: relocate a file, optionally rewriting inbound wikilinks.
+
+    Refuses moves out of OR into Sources/ and Evidence/ (append-only).
+    Curated trees on either end need `allow_curated=true`. Refuses to
+    overwrite existing destinations.
+
+    When `update_wikilinks=true` (default), scans the full vault for
+    `[[<old>]]`, `[[<old.md>]]`, and `[[<old_basename>]]` (only when the
+    basename is unique vault-wide) and rewrites them to point at the
+    new location. Preserves full-form vs stripped-form per link.
+
+    Args:
+        old_path: Vault-relative source.
+        new_path: Vault-relative destination (must not exist).
+        update_wikilinks: Default true.
+        allow_curated: Required if either end is in a curated tree.
+
+    Returns: {old_path, new_path, wikilinks_updated, files_touched, warnings}.
+    Errors: INVALID_PATH; NOT_FOUND; DEST_EXISTS; APPEND_ONLY;
+            CURATED_PROTECTED.
+    """
+    try:
+        result = move_file_module.move_file(
+            vault_root,
+            old_path=old_path,
+            new_path=new_path,
+            update_wikilinks=update_wikilinks,
+            allow_curated=allow_curated,
+        )
+    except move_file_module.MoveFileError as e:
+        raise ValueError(f"{e.code}: {e.reason}") from e
+    return result.as_dict()
+
+
+def op_delete(
+    vault_root: Path,
+    path: str,
+    confirm: bool,
+    recursive: bool = False,
+    force_orphan: bool = False,
+    force_superseded: bool = False,
+    allow_curated: bool = False,
+    expected_dead_inbound: list[str] | None = None,
+) -> dict:
+    """Tier 2: trash a file OR folder (auto-detected). Reversible — moves to
+    _trash/, not /dev/null.
+
+    Dispatches on the path: a directory is trashed whole (needs
+    `recursive=true` if non-empty; folds in the former `delete_directory`),
+    otherwise a single file. `force_superseded`/`expected_dead_inbound`
+    apply to files; `recursive` applies to folders.
+
+    Deletes are NEVER permanent at this layer. The file moves to
+    `Knowledge Base/_trash/YYYY-MM-DD/HHMMSS-<sanitized-original-path>.md`
+    with a `.meta.json` sidecar capturing original path, timestamp,
+    inbound link count, and which force-flags were used. Recovery is
+    `move_file` from the trash path back. Permanent removal happens
+    desk-side via `rm Knowledge Base/_trash/...`.
+
+    Per SKILL.md rule 6, supersession via `replace` is still preferred
+    for compiled material. Use this op for scratch, mistakes outside the
+    typed-note set, and cleanup of files that genuinely shouldn't exist.
+
+    Refuses:
+    - Sources/, Evidence/ (append-only).
+    - Files already in `_trash/` (already trashed — recover via move_file).
+    - Curated trees unless `allow_curated=true`.
+    - When `confirm=false`.
+    - When `superseded_by:` is set (history) unless `force_superseded=true`.
+    - When inbound wikilinks exist (after `expected_dead_inbound` filtering)
+      unless `force_orphan=true`.
+
+    Args:
+        path: Vault-relative.
+        confirm: Must be `true` explicitly. Marks the action deliberate.
+        recursive: For a non-empty FOLDER, required to confirm you know it
+            has contents. Ignored for files.
+        force_orphan: Allow trash even if inbound wikilinks exist.
+        force_superseded: Allow trash of a file in the supersession chain.
+        allow_curated: Required to trash under a curated tree.
+        expected_dead_inbound: Vault-relative paths whose inbound links
+            to this file should be ignored. Use when you're trashing
+            multiple files in one workflow (e.g. cleaning a supersession
+            chain) and don't want each step to false-positive on
+            links that will die in the same batch.
+
+    Returns (file): {path, trash_path, inbound_link_count,
+            inbound_ignored_count, warnings}.
+    Returns (dir): {path, trash_path, file_count, inbound_link_count,
+            warnings}.
+    Errors: UNCONFIRMED; INVALID_PATH; NOT_FOUND; ALREADY_TRASHED;
+            APPEND_ONLY; CURATED_PROTECTED; SUPERSEDED_HISTORY;
+            INBOUND_LINKS; TRASH_FAILED; (dir) NOT_A_DIR; NOT_EMPTY.
+    """
+    try:
+        abs_path, _rel = resolve_under_vault(vault_root, path)
+        is_dir = abs_path.is_dir()
+    except VaultPathError:
+        is_dir = False  # let the file backend raise the precise path error
+    try:
+        if is_dir:
+            result = delete_directory_module.delete_directory(
+                vault_root,
+                path=path,
+                confirm=confirm,
+                recursive=recursive,
+                force_orphan=force_orphan,
+                allow_curated=allow_curated,
+            )
+        else:
+            result = delete_file_module.delete_file(
+                vault_root,
+                path=path,
+                confirm=confirm,
+                force_orphan=force_orphan,
+                force_superseded=force_superseded,
+                allow_curated=allow_curated,
+                expected_dead_inbound=expected_dead_inbound,
+            )
+    except (
+        delete_file_module.DeleteFileError,
+        delete_directory_module.DeleteDirectoryError,
+    ) as e:
+        raise ValueError(f"{e.code}: {e.reason}") from e
+    return result.as_dict()
+
+
+def op_append_to_file(
+    vault_root: Path,
+    path: str,
+    content: str,
+    allow_curated: bool = False,
+) -> dict:
+    """Tier 2: append text to an existing file.
+
+    Refuses Sources/ (immutable). Allowed on Evidence/ sidecars and
+    general vault files. Curated trees need `allow_curated=true`.
+    Ensures a single newline boundary between existing tail and new
+    content.
+
+    Args:
+        path: Vault-relative.
+        content: Text to append (text only; binaries go via /upload).
+        allow_curated: Required under curated trees.
+
+    Returns: {path, bytes_appended, warnings}.
+    Errors: INVALID_APPEND; INVALID_PATH; NOT_FOUND; NOT_A_FILE;
+            APPEND_ONLY; CURATED_PROTECTED.
+    """
+    try:
+        result = append_to_file_module.append_to_file(
+            vault_root,
+            path=path,
+            content=content,
+            allow_curated=allow_curated,
+        )
+    except append_to_file_module.AppendError as e:
+        raise ValueError(f"{e.code}: {e.reason}") from e
+    return result.as_dict()
+
+
+def op_list_trash(
+    vault_root: Path,date: str | None = None) -> dict:
+    """Tier 2: enumerate recoverable trash entries. Read-only.
+
+    Walks Knowledge Base/_trash/YYYY-MM-DD/ and parses each .meta.json
+    sidecar. Returns entries most-recent-first with original path,
+    timestamp, kind (file or directory), and which force-flags fired
+    at trash time. Also surfaces drift: orphan_sidecars (sidecars with
+    no target file) and orphan_files (trashed files with no sidecar).
+    Pair with `recover_from_trash` to undo.
+
+    Args:
+        date: Optional YYYY-MM-DD filter to scope to one day.
+
+    Returns: {entries: [{trash_path, meta_path, original_path,
+             trashed_at, kind, file_count, ...}], count,
+             orphan_sidecars, orphan_files}.
+    """
+    result = list_trash_module.list_trash(vault_root, date=date)
+    return result.as_dict()
+
+
+def op_recover_from_trash(
+    vault_root: Path,
+    trash_path: str,
+    restore_path: str | None = None,
+    allow_curated: bool = False,
+) -> dict:
+    """Tier 2: undo a delete_file/delete_directory.
+
+    Reads the .meta.json sidecar to discover where the file lived
+    before being trashed, moves it back there, and cleans up the
+    sidecar. If `restore_path` is provided, uses that instead of the
+    sidecar's original location (useful when the original parent
+    directory has been removed).
+
+    Refuses to overwrite existing files at the restore destination.
+    Refuses restore into Sources/Evidence (append-only). Curated trees
+    need `allow_curated=true`.
+
+    Args:
+        trash_path: Vault-relative path to the trashed entry
+            (under `Knowledge Base/_trash/...`).
+        restore_path: Optional override; defaults to the original
+            location from the sidecar.
+        allow_curated: Required if restoring into a curated tree.
+
+    Returns: {trash_path, restored_path, kind, warnings}.
+    Errors: INVALID_PATH; NOT_FOUND; NOT_IN_TRASH; NO_RESTORE_PATH;
+            RESTORE_INTO_TRASH; APPEND_ONLY; CURATED_PROTECTED;
+            DEST_EXISTS; RECOVER_FAILED.
+    """
+    try:
+        result = recover_from_trash_module.recover_from_trash(
+            vault_root,
+            trash_path=trash_path,
+            restore_path=restore_path,
+            allow_curated=allow_curated,
+        )
+    except recover_from_trash_module.RecoverError as e:
+        raise ValueError(f"{e.code}: {e.reason}") from e
+    return result.as_dict()
+
+
+def op_list_inbound_links(
+    vault_root: Path,target: str) -> dict:
+    """Tier 2: find files whose wikilinks resolve to `target`. Read-only.
+
+    Useful before `move_file` (preview what update_wikilinks will touch)
+    or `delete_file` (preview what would break). Matches three forms:
+    - Full path: `[[Knowledge Base/Notes/Insights/foo]]`
+    - KB-stripped: `[[Notes/Insights/foo]]`
+    - Bare basename (only when unique vault-wide): `[[foo]]`
+
+    Args:
+        target: Vault-relative path or bare basename. `.md` optional.
+
+    Returns: {target, inbound: [{path, line_number, context, raw_target}],
+             count}.
+    Errors: INVALID_TARGET; INVALID_PATH.
+    """
+    try:
+        result = list_inbound_links_module.list_inbound_links(
+            vault_root, target=target
+        )
+    except list_inbound_links_module.ListInboundLinksError as e:
+        raise ValueError(f"{e.code}: {e.reason}") from e
+    return result.as_dict()
+
+
+# --------------------------------------------------------------------------- #
+# Param derivation — the declarative Param specs come straight from each leaf's
+# signature (the single source) + its docstring Args (for help text), so they can
+# never drift from what the leaf actually accepts.
+# --------------------------------------------------------------------------- #
+def _parse_args_help(doc: str | None) -> dict[str, str]:
+    """Best-effort `{param: one-line help}` from a Google-style `Args:` block."""
+    if not doc:
+        return {}
+    lines = inspect.cleandoc(doc).splitlines()
+    try:
+        start = next(i for i, ln in enumerate(lines) if ln.strip() == "Args:")
+    except StopIteration:
+        return {}
+    out: dict[str, str] = {}
+    cur: str | None = None
+    buf: list[str] = []
+    for ln in lines[start + 1:]:
+        if ln.strip() and not ln.startswith((" ", "\t")):
+            break  # next top-level section (Returns:/Errors:/…)
+        stripped = ln.strip()
+        # A new param entry looks like "name: text" with a bare identifier key.
+        head, sep, rest = stripped.partition(":")
+        if sep and head and head.replace("_", "").isalnum() and " " not in head:
+            if cur is not None:
+                out[cur] = " ".join(buf).strip()
+            cur, buf = head, [rest.strip()]
+        elif cur is not None:
+            buf.append(stripped)
+    if cur is not None:
+        out[cur] = " ".join(buf).strip()
+    return out
+
+
+def _type_tag(annotation: object) -> str:
+    """Map a (resolved) type annotation to a coercion tag."""
+    origin = typing.get_origin(annotation)
+    # Both `typing.Optional[...]`/`typing.Union[...]` and PEP-604 `X | None`
+    # (which resolves to `types.UnionType`) must be unwrapped.
+    if origin is typing.Union or origin is types.UnionType:
+        non_none = [a for a in typing.get_args(annotation) if a is not type(None)]
+        if len(non_none) == 1:
+            return _type_tag(non_none[0])
+        return "json"  # genuine multi-type union (e.g. edit's `value`)
+    if annotation is bool:
+        return "bool"
+    if annotation is int:
+        return "int"
+    if annotation is str:
+        return "str"
+    if annotation is dict or origin is dict:
+        return "dict"
+    if annotation is list or origin is list:
+        args = typing.get_args(annotation)
+        return "list[str]" if args in ((), (str,)) else "json"
+    return "json"
+
+
+def _derive_params(
+    leaf: Callable, *, skip: int, positional: str | None = None
+) -> tuple[Param, ...]:
+    """Derive the declarative `Param` tuple from a leaf signature + its docstring."""
+    sig = inspect.signature(leaf)
+    try:
+        hints = typing.get_type_hints(leaf)
+    except Exception:  # noqa: BLE001
+        hints = {}
+    helps = _parse_args_help(leaf.__doc__)
+    params: list[Param] = []
+    for p in list(sig.parameters.values())[skip:]:
+        ann = hints.get(p.name, p.annotation)
+        params.append(
+            Param(
+                name=p.name,
+                type=_type_tag(ann),
+                required=p.default is inspect.Parameter.empty,
+                help=helps.get(p.name, ""),
+                cli_positional=(p.name == positional),
+            )
+        )
+    return tuple(params)
+
+
+def note_description(project_keys_hint: str) -> str:
+    """The `note` MCP description with the live project-key hint substituted in.
+
+    `note` is a hand-registered MCP exception precisely because its description is
+    per-vault: the build injects the current project-key list/contract here so the
+    tool schema advertises live keys instead of a frozen list.
+    """
+    return (op_note.__doc__ or "").replace("__PROJECT_KEYS_HINT__", project_keys_hint)
+
+
+# --------------------------------------------------------------------------- #
+# The registry
+# --------------------------------------------------------------------------- #
+# (name, leaf, tier, cli_writes, needs_schema, cli_positional, surfaces)
+_MCRC = frozenset({"mcp", "rest", "cli"})
+_RC = frozenset({"rest", "cli"})
+_SPEC: tuple[tuple, ...] = (
+    ("find", op_find, 1, False, False, "query", _MCRC),
+    ("suggest_links", op_suggest_links, 1, False, False, None, _MCRC),
+    ("add", op_add, 1, True, True, None, _MCRC),
+    ("audit", op_audit, 1, False, False, None, _MCRC),
+    ("audit_fix", op_audit_fix, 1, True, False, None, _MCRC),
+    ("reconcile", op_reconcile, 1, True, False, None, _MCRC),
+    ("provenance_report", op_provenance_report, 1, False, False, None, _MCRC),
+    ("propose_compilation", op_propose_compilation, 1, False, False, None, _MCRC),
+    ("get", op_get, 1, False, False, "path", _MCRC),
+    ("edit", op_edit, 1, True, False, "path", _MCRC),
+    ("replace", op_replace, 1, True, False, "old_path", _MCRC),
+    ("link", op_link, 1, True, False, None, _MCRC),
+    ("preserve", op_preserve, 1, True, False, None, _MCRC),
+    # `note` is a hand-registered MCP exception (per-vault description); the registry
+    # still drives its REST route + CLI subcommand from the same leaf.
+    ("note", op_note, 1, True, False, None, _RC),
+    ("query_data", op_query_data, 2, False, False, "path", _MCRC),
+    ("create_file", op_create_file, 2, True, False, "path", _MCRC),
+    ("list_directory", op_list_directory, 2, False, False, "path", _MCRC),
+    ("move_file", op_move_file, 2, True, False, None, _MCRC),
+    ("delete", op_delete, 2, True, False, "path", _MCRC),
+    ("append_to_file", op_append_to_file, 2, True, False, "path", _MCRC),
+    ("list_trash", op_list_trash, 2, False, False, None, _MCRC),
+    ("recover_from_trash", op_recover_from_trash, 2, True, False, "trash_path", _MCRC),
+    ("list_inbound_links", op_list_inbound_links, 2, False, False, "target", _MCRC),
+)
+
+
+def _build_commands() -> tuple[Command, ...]:
+    cmds: list[Command] = []
+    for name, leaf, tier, writes, needs_schema, positional, surfaces in _SPEC:
+        skip = 2 if needs_schema else 1
+        desc = leaf.__doc__ or ""
+        if name == "note":
+            # Keep the registry description (OpenAPI/help) free of the MCP-only
+            # placeholder; the live-hint substitution happens at MCP registration.
+            desc = desc.replace("__PROJECT_KEYS_HINT__", "(any slug; unknown keys auto-register on first use)")
+        cmds.append(
+            Command(
+                name=name,
+                leaf=leaf,
+                params=_derive_params(leaf, skip=skip, positional=positional),
+                surfaces=surfaces,
+                tier=tier,
+                cli_writes=writes,
+                needs_schema=needs_schema,
+                description=desc,
+            )
+        )
+    return tuple(cmds)
+
+
+COMMANDS: tuple[Command, ...] = _build_commands()
+
+# MCP tools that are NOT produced by the generic registry loop and stay
+# hand-registered in server.py:
+#   - note: per-vault description (live project-key hint) — registered via
+#     bind_vault(op_note, …, description=note_description(hint));
+#   - mint_upload_token / mint_download_token: bound to server env (upload token,
+#     base URL), take no vault_root, and have no REST/CLI surface.
+# The fidelity test asserts every live MCP tool is either registry-generated or
+# named here — no silent gaps.
+HAND_REGISTERED_EXCEPTIONS: frozenset[str] = frozenset(
+    {"note", "mint_upload_token", "mint_download_token"}
+)
+
+
+def commands_for(surface: str, *, expose_tier2: bool = True) -> tuple[Command, ...]:
+    """The commands exposed on `surface`, honoring the tier-2 opt-out."""
+    return tuple(
+        c for c in COMMANDS if surface in c.surfaces and (expose_tier2 or c.tier == 1)
+    )
